requirements.txt
Flask==2.3.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
Flask-SocketIO==5.3.4
Flask-SocketIO==5.3.6
SQLAlchemy==2.0.23
Werkzeug==2.3.7
WTForms==3.0.1
python-dotenv==1.0.0
bcrypt==4.0.1
Pillow==10.0.1
pdfkit==1.0.0
boto3==1.28.62
google-generativeai==0.3.1
email-validator==2.0.0
python-socketio==5.9.0
eventlet==0.33.3
pymongo==4.5.0
gunicorn==21.2.0
python-magic==0.4.27

